ref: eliminate unnecessary base class inheritance for test_redis similarity backend test
